Adding a check for rownames in a column
This is in response to a problem Leo had with reading in data that he had previously written out with write.csv(..., row.names=T). I am adding a very minimal check for whether the first column should be removed, but the main answer should still be to just not have extra row identifiers.